// app/client/zod/endpoints/zheNUToolsAPI.zod.ts
export const BotControllerUpdateBotSettingsBody = zod.object({
  "answerOnFirstMessage": zod.boolean().optional().describe('Флаг, который отвечает за необходимость ответа только на первое сообщение, отправленное пользователем в чате'),
  "chatTypes": zod.array(zod.enum(['u2i', 'u2u'])).optional().describe('Типы чатов, на которые будет работать автоответ'),
  "isActive": zod.boolean().optional().describe('Флаг активности бота')
}).describe('Объект с данными для обновления настроек бота')